+++ ChangeLog.md
@@ -13,6 +13,10 @@ Other enhancements:
 
 Bug fixes:
 
+* Handle a change in GHC's hi-dump format around `addDependentFile`,
+  which now includes a hash. See
+  [yesodweb/yesod#1551](https://github.com/yesodweb/yesod/issues/1551)
+
 
 ## v1.9.1